+++ src/OpenTelemetry.AutoInstrumentation.AspNetCoreBootstrapper/BootstrapperHostingStartup.cs
@@ -15,7 +15,7 @@
 // </copyright>
 
 using Microsoft.AspNetCore.Hosting;
-using OpenTelemetry.AutoInstrumentation.Configuration;
+using OpenTelemetry.AutoInstrumentation.Configurations;
 using OpenTelemetry.AutoInstrumentation.Logger;
 
 [assembly: HostingStartup(typeof(OpenTelemetry.AutoInstrumentation.AspNetCoreBootstrapper.BootstrapperHostingStartup))]
@@ -45,13 +45,13 @@ internal class BootstrapperHostingStartup : IHostingStartup
     {
         if (!_logSettings.LogsEnabled)
         {
-            BootstrapperEventSource.Log.Trace("BootstrapperHostingStartup loaded, but OpenTelemetry Logs disabled. Skipping.");
+            Logger.LogInformation("BootstrapperHostingStartup loaded, but OpenTelemetry Logs disabled. Skipping.");
             return;
         }
 
         if (!_logSettings.EnabledInstrumentations.Contains(LogInstrumentation.ILogger))
         {
-            BootstrapperEventSource.Log.Trace($"BootstrapperHostingStartup loaded, but {nameof(LogInstrumentation.ILogger)} instrumentation is disabled. Skipping.");
+            Logger.LogInformation($"BootstrapperHostingStartup loaded, but {nameof(LogInstrumentation.ILogger)} instrumentation is disabled. Skipping.");
             return;
         }
 
@@ -60,11 +60,11 @@ internal class BootstrapperHostingStartup : IHostingStartup
             builder.ConfigureLogging(logging => logging.AddOpenTelemetryLogs());
 
             var applicationName = GetApplicationName();
-            BootstrapperEventSource.Log.Trace($"BootstrapperHostingStartup loaded for application with name {applicationName}.");
+            Logger.LogInformation($"BootstrapperHostingStartup loaded for application with name {applicationName}.");
         }
         catch (Exception ex)
         {
-            BootstrapperEventSource.Log.Error($"Error in BootstrapperHostingStartup: {ex}");
+            Logger.LogError($"Error in BootstrapperHostingStartup: {ex}");
             throw;
         }
     }
@@ -77,7 +77,7 @@ internal class BootstrapperHostingStartup : IHostingStartup
         }
         catch (Exception ex)
         {
-            BootstrapperEventSource.Log.Error($"Error getting AppDomain.CurrentDomain.FriendlyName: {ex}");
+            Logger.LogError($"Error getting AppDomain.CurrentDomain.FriendlyName: {ex}");
             return string.Empty;
         }
     }