+++ src/OpenTelemetry.AutoInstrumentation.Loader/Properties/AssemblyInfo.cs
@@ -16,4 +16,5 @@
 
 using System.Runtime.CompilerServices;
 
+[assembly: InternalsVisibleTo("Benchmarks")]
 [assembly: InternalsVisibleTo("OpenTelemetry.AutoInstrumentation.Loader.Tests")]