+++ src/OpenTelemetry.AutoInstrumentation/Instrumentation.cs
@@ -15,7 +15,7 @@
 // </copyright>
 
 using System.Diagnostics.Tracing;
-using OpenTelemetry.AutoInstrumentation.Configuration;
+using OpenTelemetry.AutoInstrumentation.Configurations;
 using OpenTelemetry.AutoInstrumentation.Diagnostics;
 using OpenTelemetry.AutoInstrumentation.Loading;
 using OpenTelemetry.AutoInstrumentation.Logging;
@@ -33,7 +33,7 @@ namespace OpenTelemetry.AutoInstrumentation;
 /// </summary>
 internal static class Instrumentation
 {
-    private static readonly ILogger Logger = OtelLogging.GetLogger();
+    private static readonly IOtelLogger Logger = OtelLogging.GetLogger();
     private static readonly LazyInstrumentationLoader LazyInstrumentationLoader = new();
 
     private static int _initialized;
@@ -69,7 +69,7 @@ internal static class Instrumentation
 
     internal static ILifespanManager LifespanManager => LazyInstrumentationLoader.LifespanManager;
 
-    internal static GeneralSettings GeneralSettings { get; } = Settings.FromDefaultSources<GeneralSettings>();
+    internal static Lazy<GeneralSettings> GeneralSettings { get; } = new(Settings.FromDefaultSources<GeneralSettings>);
 
     internal static Lazy<TracerSettings> TracerSettings { get; } = new(Settings.FromDefaultSources<TracerSettings>);
 
@@ -91,10 +91,25 @@ internal static class Instrumentation
             return;
         }
 
+#if NETFRAMEWORK
         try
         {
-            _pluginManager = new PluginManager(GeneralSettings);
+            // On .NET Framework only, initialize env vars from app.config/web.config
+            // this does not override settings which where already set via env vars.
+            // We are doing so as the OTel .NET SDK only supports the env vars and we want to be
+            // be able to set them via app.config/web.config.
+            EnvironmentInitializer.Initialize(System.Configuration.ConfigurationManager.AppSettings);
+        }
+        catch (Exception ex)
+        {
+            Logger.Error(ex, "Failed to initialize from AppSettings.");
+            throw;
+        }
+#endif
 
+        try
+        {
+            _pluginManager = new PluginManager(GeneralSettings.Value);
             _pluginManager.Initializing();
 
             if (TracerSettings.Value.TracesEnabled || MetricSettings.Value.MetricsEnabled)
@@ -106,7 +121,7 @@ internal static class Instrumentation
                 AppDomain.CurrentDomain.ProcessExit += OnExit;
                 AppDomain.CurrentDomain.DomainUnload += OnExit;
 
-                if (GeneralSettings.FlushOnUnhandledException)
+                if (GeneralSettings.Value.FlushOnUnhandledException)
                 {
                     AppDomain.CurrentDomain.UnhandledException += OnUnhandledException;
                 }
@@ -116,26 +131,43 @@ internal static class Instrumentation
 
             if (TracerSettings.Value.TracesEnabled)
             {
-                var builder = Sdk
-                    .CreateTracerProviderBuilder()
-                    .ConfigureResource(ResourceConfigurator.Configure)
-                    .UseEnvironmentVariables(LazyInstrumentationLoader, TracerSettings.Value, _pluginManager)
-                    .InvokePlugins(_pluginManager);
+                if (GeneralSettings.Value.SetupSdk)
+                {
+                    var builder = Sdk
+                        .CreateTracerProviderBuilder()
+                        .ConfigureResource(ResourceConfigurator.Configure)
+                        .UseEnvironmentVariables(LazyInstrumentationLoader, TracerSettings.Value, _pluginManager)
+                        .InvokePlugins(_pluginManager);
 
-                _tracerProvider = builder.Build();
-                Logger.Information("OpenTelemetry tracer initialized.");
+                    _tracerProvider = builder.Build();
+                    Logger.Information("OpenTelemetry tracer initialized.");
+                }
+                else
+                {
+                    AddLazilyLoadedTraceInstrumentations(LazyInstrumentationLoader, _pluginManager, TracerSettings.Value.EnabledInstrumentations);
+                    Logger.Information("Initialized lazily-loaded trace instrumentations without initializing sdk.");
+                }
             }
 
             if (MetricSettings.Value.MetricsEnabled)
             {
-                var builder = Sdk
-                    .CreateMeterProviderBuilder()
-                    .ConfigureResource(ResourceConfigurator.Configure)
-                    .UseEnvironmentVariables(LazyInstrumentationLoader, MetricSettings.Value, _pluginManager)
-                    .InvokePlugins(_pluginManager);
+                if (GeneralSettings.Value.SetupSdk)
+                {
+                    var builder = Sdk
+                        .CreateMeterProviderBuilder()
+                        .ConfigureResource(ResourceConfigurator.Configure)
+                        .UseEnvironmentVariables(LazyInstrumentationLoader, MetricSettings.Value, _pluginManager)
+                        .InvokePlugins(_pluginManager);
 
-                _meterProvider = builder.Build();
-                Logger.Information("OpenTelemetry meter initialized.");
+                    _meterProvider = builder.Build();
+                    Logger.Information("OpenTelemetry meter initialized.");
+                }
+                else
+                {
+                    AddLazilyLoadedMetricInstrumentations(LazyInstrumentationLoader, MetricSettings.Value.EnabledInstrumentations);
+
+                    Logger.Information("Initialized lazily-loaded metric instrumentations without initializing sdk.");
+                }
             }
         }
         catch (Exception ex)
@@ -147,6 +179,84 @@ internal static class Instrumentation
         if (TracerSettings.Value.OpenTracingEnabled)
         {
             EnableOpenTracing();
+        }
+    }
+
+    private static void AddLazilyLoadedMetricInstrumentations(LazyInstrumentationLoader lazyInstrumentationLoader, IList<MetricInstrumentation> enabledInstrumentations)
+    {
+        foreach (var instrumentation in enabledInstrumentations)
+        {
+            switch (instrumentation)
+            {
+                case MetricInstrumentation.AspNet:
+                    DelayedInitialization.Metrics.AddAspNet(lazyInstrumentationLoader);
+                    break;
+                case MetricInstrumentation.HttpClient:
+                    DelayedInitialization.Metrics.AddHttpClient(lazyInstrumentationLoader);
+                    break;
+                case MetricInstrumentation.NetRuntime:
+                    break;
+                case MetricInstrumentation.Process:
+                    break;
+                case MetricInstrumentation.NServiceBus:
+                    break;
+                default:
+                    Logger.Warning($"Configured metric intrumentation type is not supported: {instrumentation}");
+                    break;
+            }
+        }
+    }
+
+    private static void AddLazilyLoadedTraceInstrumentations(LazyInstrumentationLoader lazyInstrumentationLoader, PluginManager pluginManager, IList<TracerInstrumentation> enabledInstrumentations)
+    {
+        foreach (var instrumentation in enabledInstrumentations)
+        {
+            switch (instrumentation)
+            {
+                case TracerInstrumentation.AspNet:
+                    DelayedInitialization.Traces.AddAspNet(lazyInstrumentationLoader, pluginManager);
+                    break;
+                case TracerInstrumentation.HttpClient:
+                    DelayedInitialization.Traces.AddHttpClient(lazyInstrumentationLoader, pluginManager);
+                    break;
+                case TracerInstrumentation.GrpcNetClient:
+                    DelayedInitialization.Traces.AddGrpcClient(lazyInstrumentationLoader, pluginManager);
+                    break;
+                case TracerInstrumentation.SqlClient:
+                    DelayedInitialization.Traces.AddSqlClient(lazyInstrumentationLoader, pluginManager);
+                    break;
+                case TracerInstrumentation.Wcf:
+                    DelayedInitialization.Traces.AddWcf(lazyInstrumentationLoader, pluginManager);
+                    break;
+                case TracerInstrumentation.Quartz:
+                    DelayedInitialization.Traces.AddQuartz(lazyInstrumentationLoader, pluginManager);
+                    break;
+#if NET6_0_OR_GREATER
+                case TracerInstrumentation.MySqlData:
+                    DelayedInitialization.Traces.AddMySqlClient(LazyInstrumentationLoader, pluginManager);
+                    break;
+                case TracerInstrumentation.EntityFrameworkCore:
+                    DelayedInitialization.Traces.AddEntityFrameworkCore(LazyInstrumentationLoader, pluginManager);
+                    break;
+                case TracerInstrumentation.MongoDB:
+                    break;
+                case TracerInstrumentation.StackExchangeRedis:
+                    break;
+                case TracerInstrumentation.MassTransit:
+                    break;
+#endif
+                case TracerInstrumentation.GraphQL:
+                    break;
+                case TracerInstrumentation.Npgsql:
+                    break;
+                case TracerInstrumentation.NServiceBus:
+                    break;
+                case TracerInstrumentation.Elasticsearch:
+                    break;
+                default:
+                    Logger.Warning($"Configured trace intrumentation type is not supported: {instrumentation}");
+                    break;
+            }
         }
     }