+++ CHANGELOG.md
@@ -15,7 +15,10 @@ This beta release is built on top of [OpenTelemetry .NET](https://github.com/ope
 
 ### Added
 
-- Add Quartz traces instrumentation.
+- Support configuring `OTEL_*` settings using `App.config` and `Web.config`.
+- Add support for Quartz traces instrumentation.
+- Add support for EntityFrameworkCore traces instrumentations.
+- Add support for custom OpenTelemetry .NET SDK through `OTEL_DOTNET_AUTO_SETUP_SDK`.
 - Add plugin support for
   `ResourceBuilder ConfigureResource(ResourceBuilder builder)`.